Results and Checks for Login
The logic for the login and the user class are done. An actual test was not made due to error found when attempted to just test the login page. Should work if said test is done. No change of page is made if the login is done write but comments are but in place to mark all 3 possible outcomes:
-Username and password are correct
-Username exists but password wrong
-Username doesn't exist.